* Added -fp-model precies to ifort commands to avoid intel optimization problems * Removed unused parameters nhlimit and h from hfind.h * Changed particle id output in hlist_*.dat to be 0-based to follow CART convention * Changed FindHaloes call to use less than nmin particles * Changed FindHaloes to unmark particles when less than parameter nmin particles found within rhalo * removed imark option
The call to FindHaloes takes a minimum particle number as a parameter.
This is redundant with the RemoveSmall call, which has been removed.

This particle number also sets an effective minimum central particle density
for halos.  If a normal rhalo is used with a large nmin (say for finding
only large, well resolved clusters), no halos will be found since none will
have nmin particles within rhalo.

A low value passed to FindHaloes, on the other hand, will produce many
potential halo centers, which are costly to unbind.

Either a moderate nmin should be used and passed to FindHaloes, or the
parameter to FindHaloes changed to a suitably low value (20-40 particles,
say).